==> 02_encoding_ver2_hidden.ipynb <==
# mieyhgnaj set 🤪

import numpy as np
import pandas as pd

import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
sns.set_style("white")

import missingno as msno
import plotly.express as px
from matplotlib import cm
from matplotlib import style
from matplotlib import font_manager
from matplotlib import rc

rc('font', family='Arial Unicode MS')

import warnings
warnings.filterwarnings("ignore")

## 🚙 `타깃 컬럼을 제외한 컬럼은 SOCAR 측의 요청대로 비공개 처리!`

socar_raw = pd.read_csv('datas/socar_raw.csv', index_col=0)
socar_raw

# 🤪 negative values 제거!.!

# 🤪

socar = socar_raw.copy()

# Encoding
- re-Labeling
- OneHot

socar.columns

relabel = ['x6', 'x12', 'x21', 'x13', 'x23']

onehot = ['x1', 'x3', 'x11', 'x13', 'x16', 'x20']

### 1. re-labeling
- 라벨과 내용의 순서가 이어지지 않아서 re-labeling 진행

relabel

df_relabel = socar[relabel]
df_relabel

# 변경 전
df_relabel['x6'].value_counts()

# 변경 후
mapping_value = {2:1}
df_relabel = df_relabel.replace({'x6': mapping_value})
df_relabel['x6'].value_counts()

# 변경 전
df_relabel['x12'].value_counts()

# 변경 후
mapping_value = {5:0, 2:3, 3:4, 4:5}
df_relabel = df_relabel.replace({'x12': mapping_value})
df_relabel['x12'].value_counts()

# 변경 전
df_relabel['x21'].value_counts()

# 변경 후
mapping_value = {2:0}
df_relabel = df_relabel.replace({'x21': mapping_value})
df_relabel['x21'].value_counts()

# 변경 전
df_relabel['x13'].value_counts()

# 변경 후
mapping_value = {-1:0}
df_relabel = df_relabel.replace({'x13': mapping_value})
df_relabel['x13'].value_counts()

# 변경 전
df_relabel['x23'].value_counts()

# 변경 후
mapping_value = {-1:0, 0:1, 1:2, 2:3, 3:4, 4:5}
df_relabel = df_relabel.replace({'x23': mapping_value})
df_relabel['x23'].value_counts()

df_relabel

socar_relabel = socar.copy()

# 데이터 교체하기
socar_relabel = socar_relabel.drop(columns=relabel)
socar_relabel = pd.concat((socar_relabel, df_relabel), axis=1)

socar_relabel

#### train, test 분리

train = socar_relabel[socar_relabel['x24']==0]
train.drop(columns=['x24'], inplace=True)
test = socar_relabel[socar_relabel['x24']==1]
test.drop(columns=['x24'], inplace=True)

train.reset_index(drop=True, inplace=True)
test.reset_index(drop=True, inplace=True)

train

test

### 2. OneHotEncoder()
- 기존에 라벨링 되어있는 변수들 중에 순서가 없는 변수는 원핫인코딩 진행

# 원핫인코딩할 변수
onehot

# train

df_onehot_train = train[onehot]
df_onehot_train

# 33개 생긴다

from sklearn.preprocessing import OneHotEncoder

enc = OneHotEncoder(handle_unknown='ignore', sparse=False)
enc.fit(df_onehot_train)
df_onehot_train = pd.DataFrame(data=enc.transform(df_onehot_train), columns=enc.get_feature_names(onehot), dtype=bool)
# trasfer true and false to 1 and 0
df_onehot_train = df_onehot_train * 1

df_onehot_train

df_onehot_train.columns, len(df_onehot_train.columns)

# 데이터 합치기
train_encode = pd.concat((train, df_onehot_train), axis=1)
train_encode = train_encode.drop(columns=onehot)

train_encode.columns, len(train_encode.columns)

# test

df_onehot_test = test[onehot]
df_onehot_test

df_onehot_test = pd.DataFrame(data=enc.transform(df_onehot_test), columns=enc.get_feature_names(onehot), dtype=bool)
df_onehot_test = df_onehot_test * 1

df_onehot_test

# 데이터 합치기
test_encode = pd.concat((test, df_onehot_test), axis=1)
test_encode = test_encode.drop(columns=onehot)

test_encode.columns, len(test_encode.columns)

# re-label, OneHot 완료
train_encode.to_csv('datas/train_encode_ver2.csv')
test_encode.to_csv('datas/test_encode_ver2.csv')

# 불러올 때🤪

train = pd.read_csv('datas/train_encode_ver2.csv', index_col=0)
test = pd.read_csv('datas/test_encode_ver2.csv', index_col=0)

# 나눌땐 이렇게 나눠서 쓰면 되지롱
X_train = train.drop(columns=['fraud_YN'])
X_test = test.drop(columns=['fraud_YN'])
y_train = train.fraud_YN
y_test = test.fraud_YN
