==> 03_oversampling_hidden.ipynb <==
# mieyhgnaj set 🤪

import numpy as np
import pandas as pd

import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
sns.set_style("white")

import missingno as msno
import plotly.express as px
from matplotlib import cm
from matplotlib import style
from matplotlib import font_manager
from matplotlib import rc

rc('font', family='Arial Unicode MS')

import warnings
warnings.filterwarnings("ignore")

## 🚙 `타깃 컬럼을 제외한 컬럼은 SOCAR 측의 요청대로 비공개 처리!`

### 0. encoding ver.

train = pd.read_csv('datas/train_encode_ver2.csv', index_col=0)
test = pd.read_csv('datas/test_encode_ver2.csv', index_col=0)

X_train = train.drop(columns=['fraud_YN'])
X_test = test.drop(columns=['fraud_YN'])
y_train = train.fraud_YN
y_test = test.fraud_YN

X_train.shape, X_test.shape, y_train.shape, y_test.shape

y_train.value_counts()

# re-sampling
- Oversampling: Duplicating samples from the minority class
- Undersampling: Deleting samples from the majority class

### 1. Naive random over-sampling
: generate new samples by randomly sampling with replacement the current available samples

from collections import Counter

from imblearn.over_sampling import RandomOverSampler

random = RandomOverSampler(random_state=4)
X_resampled, y_resampled = random.fit_resample(X_train, y_train)

print(random.__class__.__name__, "\n")
print(sorted(Counter(y_resampled).items()))

print(X_resampled[-5:])

y_resampled.value_counts()

### 2. SMOTE
: First it finds the n-nearest neighbors in the minority class for each of the samples in the class. Then it draws a line between the the neighbors an generates random points on the lines.

from imblearn.over_sampling import SMOTE

smote = SMOTE()
X_resampled, y_resampled = smote.fit_resample(X_train, y_train)
print(sorted(Counter(y_resampled).items()))

### 3. ADASYN
: Improved version of Smote.

from imblearn.over_sampling import ADASYN

adasyn = ADASYN()
X_resampled, y_resampled = adasyn.fit_resample(X_train, y_train)

print(adasyn.__class__.__name__, "\n")
print(sorted(Counter(y_resampled).items()))

print(X_resampled[-5:])

y_resampled.value_counts()

### 4. BorderlineSMOTE
: Unlike with the SMOTE, where the synthetic data are created randomly between the two data, Borderline-SMOTE only makes synthetic data along the decision boundary between the two classes.

from imblearn.over_sampling import BorderlineSMOTE

border = BorderlineSMOTE()
X_resampled, y_resampled = border.fit_resample(X_train, y_train)
print(sorted(Counter(y_resampled).items()))

print(X_resampled[-10:])

y_resampled.value_counts()

### 5. SMOTENC
: Synthetic Minority Over-sampling Technique for Nominal and Continuous.\
: Unlike SMOTE, SMOTE-NC for dataset containing numerical and categorical features. However, it is not designed to work with only categorical features.

# categorical_features = []
# for i,j in enumerate(list(X_train.columns)):
#     categorical_features.append(i)

# categorical_features.remove(7) # 14

# categorical_features.remove(8) # 15

categorical_features = [0, 1, 2, 3, 4, 5, 6,
                        9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26,
                        27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44,
                        45, 46, 47, 48, 49]

from imblearn.over_sampling import SMOTENC

somtenc = SMOTENC(categorical_features=categorical_features,
                                   random_state=4)
X_resampled, y_resampled = somtenc.fit_resample(X_train, y_train)

print(somtenc.__class__.__name__, "\n")
print(sorted(Counter(y_resampled).items()))

print(X_resampled[-5:])

y_resampled.value_counts()

### 6. SVMSMOTE
: What special about Borderline-SMOTE SVM compared to the Borderline-SMOTE is that more data are synthesized away from the region of class overlap. It focuses more on where the data is separated.

from imblearn.over_sampling import SVMSMOTE

svmsmote = SVMSMOTE(random_state=4)
X_resampled, y_resampled = svmsmote.fit_resample(X_train, y_train)
print(sorted(Counter(y_resampled).items()))

### 7. SMOTEENN (combination(over-under))
: Over-sampling using SMOTE and cleaning using ENN.\
: Combine over- and under-sampling using SMOTE and Edited Nearest Neighbours.

from imblearn.combine import SMOTEENN

smoteenn = SMOTEENN(random_state=4)
X_resampled, y_resampled = smoteenn.fit_resample(X_train, y_train)

print(smoteenn.__class__.__name__, "\n")
print(sorted(Counter(y_resampled).items()))

print(X_resampled[-10:])

y_resampled.value_counts()

### 8. SMOTETomek (combination(over-under))
: Over-sampling using SMOTE and cleaning using Tomek links.\
: Combine over- and under-sampling using SMOTE and Tomek links.

from imblearn.combine import SMOTETomek

smotetomek = SMOTETomek(random_state=4)
X_resampled, y_resampled = smotetomek.fit_resample(X_train, y_train)
print(sorted(Counter(y_resampled).items()))

y_resampled.value_counts()
