==> 05_outlier_removal_hidden.ipynb <==
# mieyhgnaj set 🤪

import numpy as np
import pandas as pd

import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
sns.set_style("white")

import missingno as msno
import plotly.express as px
from matplotlib import cm
from matplotlib import style
from matplotlib import font_manager
from matplotlib import rc

rc('font', family='Arial Unicode MS')

import warnings
warnings.filterwarnings("ignore")

## 🚙 `타깃 컬럼을 제외한 컬럼은 SOCAR 측의 요청대로 비공개 처리!`

train = pd.read_csv('datas/train_encode_ver2.csv', index_col=0)
test = pd.read_csv('datas/test_encode_ver2.csv', index_col=0)

# Outlier Removal

pd.set_option('display.float_format', lambda x: '%.5f' % x)

rm_out = train[['x14', 'x15']]
rm_out

data = rm_out['x15']
data

### 1. Standard Deviation Method

# identify outliers with standard deviation
from numpy import mean
from numpy import std
# calculate summary statistics
data_mean, data_std = mean(data), std(data)
# identify outliers
cut_off = data_std * 3
lower, upper = data_mean - cut_off, data_mean + cut_off
# identify outliers
outliers = [x for x in data if x < lower or x > upper]
print('Identified outliers: %d' % len(outliers))
# remove outliers
outliers_removed = [x for x in data if x >= lower and x <= upper]
print('Non-outlier observations: %d' % len(outliers_removed))

### 2. Interquartile Range Method

# identify outliers with interquartile range
from numpy import percentile
# calculate interquartile range
q25, q75 = percentile(data, 25), percentile(data, 75)
iqr = q75 - q25
print('Percentiles: 25th=%.3f, 75th=%.3f, IQR=%.3f' % (q25, q75, iqr))
# calculate the outlier cutoff
cut_off = iqr * 1.5
lower, upper = q25 - cut_off, q75 + cut_off
# identify outliers
outliers = [x for x in data if x < lower or x > upper]
print('Identified outliers: %d' % len(outliers))
# remove outliers
outliers_removed = [x for x in data if x >= lower and x <= upper]
print('Non-outlier observations: %d' % len(outliers_removed))

data.describe()

# Standard Deviation Method로 해봐야겠다

mean(outliers)

mean(outliers_removed)

# identify outliers with standard deviation
from numpy import mean
from numpy import std

def rm_outlier(data):
    # calculate summary statistics
    data_mean, data_std = mean(data), std(data)
    # identify outliers
    cut_off = data_std * 3
    lower, upper = data_mean - cut_off, data_mean + cut_off

    rm_out_data = pd.DataFrame(data)
    for idx, x in enumerate(data):
        if x < lower or x > upper:
            rm_out_data.loc[idx] = 0
#             rm_out_data.drop(idx, inplace=True)

    # identify outliers
    outliers = [x for x in data if x < lower or x > upper]
    print('Identified outliers: %d' % len(outliers))
    # remove outliers
    outliers_removed = [x for x in data if x >= lower and x <= upper]
    print('Non-outlier observations: %d' % len(outliers_removed))
    
    return rm_out_data

cost_mean['x14'].replace(0, statistics.mean(cost_mean['x14']), inplace=True)

mean(train['x15'])

rm_x15 = rm_outlier(rm_out['x15'])
rm_x15

rm_x14 = rm_outlier(rm_out['x14'])
rm_x14

### outlier 제거한 버전으로 교체

train_out = train.copy()

train_out['x15'] = rm_x15
train_out['x14'] = rm_x14
train_out

train_out.to_csv('datas/train_out.csv')
